toolkit/content/tests/widgets/test_moz_input_color.html

<!doctype html>
<html>
  <head>
    <meta charset="utf-8" />
    <title>moz-input-color Tests</title>
    <script src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
    <script
      type="module"
      src="chrome://global/content/elements/moz-input-color.mjs"
    ></script>
    <link
      rel="stylesheet"
      href="chrome://mochikit/content/tests/SimpleTest/test.css"
    />
  </head>

  <body>
    <p id="display"></p>
    <div id="content">
      <moz-input-color
        value="#ffffff"
        name="test-prop"
        data-l10n-id="color-input-test-label"
      ></moz-input-color>
    </div>
    <script src="lit-test-helpers.js"></script>
    <script class="testbody" type="application/javascript">
      let testHelpers = new InputTestHelpers();

      add_task(async function testMozInputColor() {
        const colorInput = document.querySelector("moz-input-color");

        ok(colorInput, "color input element is rendered");

        let input = colorInput.shadowRoot.querySelector("input");
        is(input.value, "#ffffff", "color input has the correct initial value");
      });
      add_task(async function testMozInputColorEvents() {
        const colorInput = document.querySelector("moz-input-color");
        let input = colorInput.shadowRoot.querySelector("input");
        let { trackEvent, verifyEvents } = testHelpers.getInputEventHelpers();

        colorInput.addEventListener("change", trackEvent);

        // Simulate a value change on the inner input
        const NEW_COLOR = "#000000";
        input.value = NEW_COLOR;
        input.dispatchEvent(new Event("change"));

        // Verify that the value change propagates to moz-input-color
        verifyEvents([
          { type: "change", localName: "moz-input-color", value: NEW_COLOR },
        ]);
        is(colorInput.value, NEW_COLOR, "Input value was updated.");
      });
    </script>
  </body>
</html>
